Inheritance of enabled & multi dimension support added
Reviewed-on: https://git.eclipse.org/r/84346
Tested-by: Hudson CI

+++ org.eclipse.scout.rt.client/src/main/java/org/eclipse/scout/rt/client/ui/basic/table/organizer/OrganizeColumnsForm.java
@@ -181,7 +181,7 @@ public class OrganizeColumnsForm extends AbstractForm implements IOrganizeColumn
           }
 
           @Override
-          protected int getConfiguredLabelPosition() {
+          protected byte getConfiguredLabelPosition() {
             return LABEL_POSITION_TOP;
           }
 
@@ -597,7 +597,7 @@ public class OrganizeColumnsForm extends AbstractForm implements IOrganizeColumn
           }
 
           @Override
-          protected int getConfiguredLabelPosition() {
+          protected byte getConfiguredLabelPosition() {
             return LABEL_POSITION_TOP;
           }
