+++ org.eclipse.scout.rt.client/src/main/java/org/eclipse/scout/rt/client/ui/desktop/bookmark/view/BookmarkViewForm.java
@@ -147,7 +147,7 @@ public class BookmarkViewForm extends AbstractForm {
           }
 
           @Override
-          protected int getConfiguredLabelPosition() {
+          protected byte getConfiguredLabelPosition() {
             return IFormField.LABEL_POSITION_ON_FIELD;
           }
